+++ mqttany/logger.py
@@ -36,14 +36,33 @@ import colorlog
 
 TRACE = 5  # trace level logging
 
-_log_file = os.path.join(os.path.dirname(__file__), "log", "mqttany.log")
-_log_format = "%(asctime)s [%(levelname)-5s] [%(name)-24s] %(message)s"
-_log_format_debug = (
-    "%(asctime)s [%(levelname)-5s] [%(processName)-8s] [%(name)-24s] %(message)s"
+_LOG_FILE = os.path.join(os.path.dirname(__file__), "log", "mqttany.log")
+
+_LOG_LEN_LEVEL = 5
+_LOG_LEN_PROCESS = 8
+_LOG_LEN_NAME = 20
+_LOG_FORMAT_FILE = (
+    "%(asctime)s [%(levelname)-{level}s] [%(name)-{name}s] %(message)s".format(
+        level=_LOG_LEN_LEVEL, name=_LOG_LEN_NAME
+    )
 )
-_log_format_color = "%(asctime)s [%(log_color)s%(levelname)-5s%(reset)s] [%(name)-24s] %(message_log_color)s%(message)s"
-_log_format_color_debug = "%(asctime)s [%(log_color)s%(levelname)-5s%(reset)s] [%(processName)-8s] [%(name)-24s] %(message_log_color)s%(message)s"
-_log_colors = {
+_LOG_FORMAT_FILE_DEBUG = (
+    "%(asctime)s [%(levelname)-{level}s] [%(processName)-{process}.{process}s] "
+    "[%(name)-{name}s] %(message)s".format(
+        level=_LOG_LEN_LEVEL, process=_LOG_LEN_PROCESS, name=_LOG_LEN_NAME
+    )
+)
+_LOG_FORMAT_TERM = (
+    "%(asctime)s [%(log_color)s%(levelname)-{level}s%(reset)s] [%(name)-{name}s] "
+    "%(message_log_color)s%(message)s".format(level=_LOG_LEN_LEVEL, name=_LOG_LEN_NAME)
+)
+_LOG_FORMAT_TERM_DEBUG = (
+    "%(asctime)s [%(log_color)s%(levelname)-{level}s%(reset)s] [%(processName)-{process}.{process}s] "
+    "[%(name)-{name}s] %(message_log_color)s%(message)s".format(
+        level=_LOG_LEN_LEVEL, process=_LOG_LEN_PROCESS, name=_LOG_LEN_NAME
+    )
+)
+_LOG_COLORS = {
     "TRACE": "white",
     "DEBUG": "bold_white",
     "INFO": "bold_green",
@@ -51,7 +70,7 @@ _log_colors = {
     "ERROR": "bold_red",
     "CRITICAL": "bold_red",
 }
-_secondary_log_colors = {
+_LOG_SECONDARY_COLORS = {
     "message": {
         "DEBUG": "bold",
         "INFO": "bold_green",
@@ -60,6 +79,8 @@ _secondary_log_colors = {
         "CRITICAL": "bold_red",
     }
 }
+
+_log_handlers = []
 
 
 def trace_log(self, msg, *args, **kwargs):
@@ -91,53 +112,53 @@ def _init_logger():
     log.warn = MethodType(warn_log, log)
     log.setLevel(INFO)
 
-    if not os.path.exists(os.path.dirname(_log_file)):
+    if not os.path.exists(os.path.dirname(_LOG_FILE)):
         try:
-            os.makedirs(os.path.dirname(_log_file))
+            os.makedirs(os.path.dirname(_LOG_FILE))
         except OSError as err:
             if err.errno != errno.EEXIST:
                 raise
     handler_file = handlers.RotatingFileHandler(
-        _log_file, maxBytes=5242880, backupCount=5
+        _LOG_FILE, maxBytes=5242880, backupCount=5
     )
-    handler_file.setFormatter(logging.Formatter(_log_format))
-    log.addHandler(handler_file)
+    handler_file.setFormatter(logging.Formatter(_LOG_FORMAT_FILE))
+    _log_handlers.append(handler_file)
 
     handler_stdout = logging.StreamHandler()
     handler_stdout.setFormatter(
         colorlog.ColoredFormatter(
-            _log_format_color,
-            log_colors=_log_colors,
-            secondary_log_colors=_secondary_log_colors,
+            _LOG_FORMAT_TERM,
+            log_colors=_LOG_COLORS,
+            secondary_log_colors=_LOG_SECONDARY_COLORS,
         )
     )
-    log.addHandler(handler_stdout)
+    _log_handlers.append(handler_stdout)
 
+    for handler in _log_handlers:
+        log.addHandler(handler)
     return log
 
 
 _log = _init_logger()
 
 
-def get_logger(name="mqttany", level=None):
+def get_logger(name=None, level=None):
     """
-    Returns a logger
+    Returns a logger, if ``name`` is not specified the import name of the calling
+    file will be used.
     """
+    if name is None:
+        frm = inspect.stack()[1]
+        name = inspect.getmodule(frm[0]).__name__
+    name = ".".join([s for s in name.split(".") if s not in ["mqttany", "modules"]])
+    name = name[len(name) - _LOG_LEN_NAME :] if len(name) > _LOG_LEN_NAME else name
     logger = logging.getLogger(name)
+    for handler in _log_handlers:
+        logger.addHandler(handler)
     logger.trace = MethodType(trace_log, logger)
     logger.warn = MethodType(warn_log, logger)
     logger.setLevel(level or logging.getLogger("mqttany").level)
     return logger
-
-
-def get_module_logger(module=None, level=None):
-    """
-    Returns a logger for a module
-    """
-    if module is None:
-        frm = inspect.stack()[1]
-        module = ".".join(inspect.getmodule(frm[0]).__name__.split(".")[1:])
-    return get_logger("mqttany.{}".format(module), level)
 
 
 def set_level(level):
@@ -146,29 +167,31 @@ def set_level(level):
     """
     _log.setLevel(level)
     if level <= DEBUG:
-        for handler in _log.handlers:
+        # for handler in _log.handlers:
+        for handler in _log_handlers:
             if isinstance(handler, logging.StreamHandler):
                 handler.setFormatter(
                     colorlog.ColoredFormatter(
-                        _log_format_color_debug,
-                        log_colors=_log_colors,
-                        secondary_log_colors=_secondary_log_colors,
+                        _LOG_FORMAT_TERM_DEBUG,
+                        log_colors=_LOG_COLORS,
+                        secondary_log_colors=_LOG_SECONDARY_COLORS,
                     )
                 )
             else:
-                handler.setFormatter(logging.Formatter(_log_format_debug))
+                handler.setFormatter(logging.Formatter(_LOG_FORMAT_FILE_DEBUG))
     else:
-        for handler in _log.handlers:
+        # for handler in _log.handlers:
+        for handler in _log_handlers:
             if isinstance(handler, logging.StreamHandler):
                 handler.setFormatter(
                     colorlog.ColoredFormatter(
-                        _log_format_color,
-                        log_colors=_log_colors,
-                        secondary_log_colors=_secondary_log_colors,
+                        _LOG_FORMAT_TERM,
+                        log_colors=_LOG_COLORS,
+                        secondary_log_colors=_LOG_SECONDARY_COLORS,
                     )
                 )
             else:
-                handler.setFormatter(logging.Formatter(_log_format))
+                handler.setFormatter(logging.Formatter(_LOG_FORMAT_FILE))
 
 
 def log_traceback(log, limit=None):
@@ -184,6 +207,7 @@ def uninit():
     """
     Flushes and closes all log handlers
     """
-    for handler in _log.handlers:
+    # for handler in _log.handlers:
+    for handler in _log_handlers:
         handler.flush()
         handler.close()